Improved support for UInt64 (although Delphi JSON framework has not UInt64 support) More tests

+++ Demos/Main/Demo.Forms.Serialization.Simple.pas
@@ -112,7 +112,7 @@ end;
 
 procedure TfrmSerializationSimple.actDesIntegerExecute(Sender: TObject);
 begin
-  DeserializeSimple<Integer>;
+  DeserializeSimple<UInt64>;
 end;
 
 procedure TfrmSerializationSimple.actDesStringExecute(Sender: TObject);
